csa: core implementation
csa enables allocation of large chunks of
continuous physical memory space.
The API is in csa.h:
- csa_alloc(): allocate the given amount on current node
- csa_alloc_node(): same for the given node id
Allocation request arguments include size (bytes) and
alignment (bytes). The later is optional, 0 yields
default alignement of a top order page.

Approaches:
1) Trivial: use alloc_pages_nodemask(,..., MAX_ORDER-1).
Keep allocating until the pages we got coalesce into the
continuous space that matches the request.
Since allocations come from the free_list, where the order
is random, especially on the top, we may run under the zone
watermark before we get what we need.
Free'ing pages would put them back onto the top of the list.
Not clear when a certain range should be abandonded and freed.

2) Walk the suitable zone free_lists and find the continuous
range assuming the large zones (DMA32/Moveable) have most of
the system memory intact and continuous as freed from bootmem
blocks.
Walk the lists backwards to mine from the bottom of the lists.
When a range is found, take if off the list and adjust
the accounting.

3) Use alloc_contig_range()/free_contig_range() implemented
in page_alloc.c, just not exported. Requires a kernel
change to export these two.
This method is used in hugetlb.c and does not care to
handle larger ranges.
First we walk the zone pages raw and find a range of free ones.
Isolation process used by alloc_contig_range generally
can move allocated pages to other pages in the moveable
zone and free the pages we need.

Option (2) is implemented and enabled when
USE_CONTIG_ALLOC is disabled (csa_core.c).
Option (3) is implemented and enabled when
USE_CONTIG_ALLOC is enabled.

Notes:
- csa.ko does not implement allocation tracking.
It is the caller's responsibility to free the space allocated.

Signed-off-by: Anton <anton.eidelman@gmail.com>